[verus] Verification failed for frame (verified:0, errors:0)

+++ verus/frame.rs
@@ -919,7 +919,7 @@ impl FrameAllocator {
                             assert(self@.is_allocated(idx as int));
                             // idx is in the range [start_frame, start_frame + count).
                             assert(start_frame as int <= idx as int);
-                            assert(idx as int < start_frame as int + count as int);
+                            assert(idx as int < (start_frame + count) as int);
                             // So there EXISTS an i in range with is_allocated(i).
                             // This is a counterexample to the forall, making the antecedent false.
                         }
